[#97] feat: inputpair 컴포넌트 스타일추가/ 폴더별로 action.ts보관/에러표시
